ONE LAST IDEA i swear.
we 1) ask github CI/CD to setup a python3.10
we 2) ask conda to create an env with a python, but the package is NOT installed.

i tried "pip install .", but that didnt work... because it used the first python, i imagine.

now i will try calling the explicit pip of the CONDA python to install the package, before calling the test.